+++ CHANGELOG.md
@@ -27,6 +27,10 @@ Thank you to all who have contributed!
 
 ### Added
 - Added SqlDialect support for MAP expression and type serialization
+- Added MAP functions: map_contains_key, map_get, map_keys, map_values, map_entries, size, exists, cardinality and `IS MAP<K, V>
+- Added implicit cast for `MAP` lookup key type if needed for bracket notation, map_contains_key function and map_get function.
+- Added MAP type operation support: CAST and UNPIVOT
+- Added `MAP` type parsing in DatumIonReader for CLI and Conformance tests
 
 ### Changed
 
@@ -40,6 +44,7 @@ Thank you to all who have contributed!
 
 ### Contributors
 Thank you to all who have contributed!
+- @xd1313113
 
 ## [1.4.0](https://github.com/partiql/partiql-lang-kotlin/releases/tag/v1.4.0) - 2026-06-05